requirements.txt
openai
streamlit
librosa
numpy
scikit-learn
matplotlib
python-dotenv
tqdm